// chpqa-app/src/app/api-services/chpqa-api/generated/model/replyScheme.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Policy } from './policy';
import { RequestEconomicSubSector } from './requestEconomicSubSector';
import { Organisation } from './organisation';
import { RequestEconomicSector } from './requestEconomicSector';
import { ReplySchemeLocationDetails } from './replySchemeLocationDetails';
import { Person } from './person';
import { SicCode } from './sicCode';


export interface ReplyScheme { 
    id?: string | null;
    name?: string | null;
    reference?: string | null;
    createdon?: string | null;
    status?: number | null;
    latestSubmissionStatus?: number | null;
    statusText?: string | null;
    company?: Organisation;
    responsiblePerson?: Person;
    site?: ReplySchemeLocationDetails;
    policyRocs?: Policy;
    policyChpqa?: Policy;
    econSector?: RequestEconomicSector;
    econSubSector?: RequestEconomicSubSector;
    sicCode?: SicCode;
}